# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Email Marketing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: groupId
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: Paginated contacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactList"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /contacts/bulk-delete:
    post:
      operationId: bulkDeleteContacts
      tags: [contacts]
      summary: Bulk delete contacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteInput"
      responses:
        "200":
          description: Deleted count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkResult"

  # ── Contact Groups ─────────────────────────────────────────────────────────
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List contact groups
      responses:
        "200":
          description: Groups list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactGroup"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a contact group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactGroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactGroup"
  /groups/{id}:
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update a group
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactGroupUpdate"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactGroup"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── Templates ──────────────────────────────────────────────────────────────
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, scheduled, running, paused, completed, cancelled]
      responses:
        "200":
          description: Campaigns list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/status:
    patch:
      operationId: updateCampaignStatus
      tags: [campaigns]
      summary: Update campaign status (start, pause, cancel)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  # ── Email Logs ─────────────────────────────────────────────────────────────
  /logs:
    get:
      operationId: listEmailLogs
      tags: [logs]
      summary: List email logs
      parameters:
        - name: campaignId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema:
            type: string
            enum: [sent, failed, pending]
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: Email logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailLogList"
    delete:
      operationId: clearEmailLogs
      tags: [logs]
      summary: Clear all email logs
      responses:
        "204":
          description: Cleared
  /logs/{id}:
    delete:
      operationId: deleteEmailLog
      tags: [logs]
      summary: Delete an email log
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── Reports ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Contact:
      type: object
      required: [id, email, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: ["string", "null"] }
        email: { type: string }
        company: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        groupIds:
          type: array
          items: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    ContactInput:
      type: object
      required: [email]
      properties:
        name: { type: string }
        email: { type: string }
        company: { type: string }
        phone: { type: string }
        website: { type: string }
        city: { type: string }
        country: { type: string }
        notes: { type: string }
        tags:
          type: array
          items: { type: string }
        groupIds:
          type: array
          items: { type: string }
    ContactUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        company: { type: string }
        phone: { type: string }
        website: { type: string }
        city: { type: string }
        country: { type: string }
        notes: { type: string }
        tags:
          type: array
          items: { type: string }
        groupIds:
          type: array
          items: { type: string }
        isActive: { type: boolean }
    ContactList:
      type: object
      required: [contacts, total, page, limit]
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    BulkDeleteInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items: { type: string }
    BulkResult:
      type: object
      required: [deleted]
      properties:
        deleted: { type: integer }

    # ── Contact Groups ────────────────────────────────────────────────────────
    ContactGroup:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        contactCount: { type: integer }
        createdAt: { type: string }
    ContactGroupInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
    ContactGroupUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }

    # ── Templates ─────────────────────────────────────────────────────────────
...
    Campaign:
      type: object
      required: [id, name, subject, senderName, senderEmail, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        subject: { type: string }
        senderName: { type: string }
        senderEmail: { type: string }
        smtpAccountId: { type: ["string", "null"] }
        templateId: { type: ["string", "null"] }
        groupId: { type: ["string", "null"] }
        status:
          type: string
          enum: [draft, scheduled, running, paused, completed, cancelled]
        scheduledAt: { type: ["string", "null"] }
        hourlyLimit: { type: ["integer", "null"] }
        dailyLimit: { type: ["integer", "null"] }
        totalRecipients: { type: integer }
        sentCount: { type: integer }
        failedCount: { type: integer }
        mailsPerBatch: { type: integer }
        intervalMinutes: { type: integer }
        customHtml: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    CampaignInput:
      type: object
      required: [name, subject, senderName, senderEmail]
      properties:
        name: { type: string }
        subject: { type: string }
        senderName: { type: string }
        senderEmail: { type: string }
        smtpAccountId: { type: string }
        templateId: { type: string }
        groupId: { type: string }
        scheduledAt: { type: string }
        hourlyLimit: { type: integer }
        dailyLimit: { type: integer }
        mailsPerBatch: { type: integer }
        intervalMinutes: { type: integer }
        customHtml: { type: string }
    CampaignUpdate:
      type: object
      properties:
        name: { type: string }
        subject: { type: string }
        senderName: { type: string }
        senderEmail: { type: string }
        smtpAccountId: { type: string }
        templateId: { type: string }
        groupId: { type: string }
        scheduledAt: { type: string }
        hourlyLimit: { type: integer }
        dailyLimit: { type: integer }
        mailsPerBatch: { type: integer }
        intervalMinutes: { type: integer }
        customHtml: { type: string }
    CampaignStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [running, paused, cancelled]

    # ── Logs ──────────────────────────────────────────────────────────────────
    EmailLog:
      type: object
      required: [id, recipient, campaignId, status, createdAt]
      properties:
        id: { type: string }
        recipient: { type: string }
        campaignId: { type: string }
        campaignName: { type: ["string", "null"] }
        smtpAccountId: { type: ["string", "null"] }
        status:
          type: string
          enum: [sent, failed, pending]
        retryCount: { type: integer }
        smtpResponse: { type: ["string", "null"] }
        error: { type: ["string", "null"] }
        createdAt: { type: string }
    EmailLogList:
      type: object
      required: [logs, total, page, limit]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/EmailLog"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── Reports ───────────────────────────────────────────────────────────────
...